move impedance based model to new file
